requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
flask-wtf
email-validator